Fixed bug where calling rtc in shell fucked fish/pingpong + added Ctrl C
Bug: executing rtc fucks shit up afterwards.  Why" Execute compared the
first 4 bytes of rtc file to the magic exe nums, found it wasn't an
executable and returned! HAVE TO CLOSE THE FILE FIRST FUUUUUUUU. Just
moved the file close up so that rtc gets set back to not in use when
this executable test fails - as well as anything similar.

Also added a jank ass ctrl c. Had to include syscalls.h in the
terminal.c file (terminal.h made compiler errors). I feel dirty fucking
around with all these includes...but it works....Anyway, keyboard
handler checks for the Ctrl C case, clears the screen, and then calls an
exception halt to end the program. Not sure if this fucks any of our
other shit up, especially terminal-wise. Michael, how does it look to
you?