Fixed weird "hello"/keyboard-terminal write miscommunication bug
The basics boils down to the error being result of bad
communication/handling of terminal-write and keyboard inputs being
possible to screen. I think with some work, there is probably a better
handling of these two inputs.

Anyway, this fix redefines SHELL_OFFSET to SHELL_PROMPT_OFFSET and
leaves it around only so that the screen can be cleared (when 391OS> is
printed by us and we know the offset should be 7).

Otherwise, a local global variable shell offset is defined. This
variable is updated with num_bytes written in terminal write -
basically, it dynamically updates with how much the user program wrote
to the beginning of the line (I can already think of a strange edge case
- for instance, the user program calls a sys call to write twice
separately such that the shell_offset is updated to be only the second
write and not the total write length).

After that any SHELL_OFFSET not dealing with clearing the screen was
replaced by this global variable.

I think a better more robust way would to be to reevaluate the way the
keyboard and terminal writes are handled such that they have some sort
of better synchronization.

Also, I tried to do shell_offset = buf_start.pos_x instead (figuring
that if multiple terminal writes were called, it would then updated to
the true offset into the line written by the user program). However,
this fucked up the cursor on the second line and lead to funky shit
instead of fixing like num_bytes does for now.