Made fileops all static (STYLE PTS BBY) and added some safety to filesys.c to be more safe cuz I liek safe, I am scared, hold me don't break
I just wanted to make it explicit that filesys name comparing stuff was
looking for a null byte as the end of file names or the max length.
Shouldn't change/affect anything.

I also made dir_read a little differently - such that it doesn't just
return 32 bytes read from the file_name field but rather reads the
file_name and null terminates it if possible.